POLIO_1982: code review
- add tests
- split test file
- optimize subactivity DB query
